201835457 박재민/src/requirements.txt
nltk==3.9.1
tabulate==0.9.0
pandas==2.2.3
PyPDF2==3.0.1
python-docx==1.1.2
transformers==4.46.3
torch==2.5.1